+++ frnk/core/util/src/commonMain/kotlin/dev/jdgarita/frnk/utils/Frnk.kt
@@ -1,5 +1,5 @@
 package dev.jdgarita.frnk.utils
 
 object Frnk {
-    const val VERSION = "0.3.1"
+    const val VERSION = "0.4.0"
 }